footswitch: Avoid calling back core from separate thread
The GPIO callback is coming from a separate thread maintained by the
GPIO library. Calling back into the rest of our code from it is racy
and will cause problem.

This reworks the footswitch code to avoid this.

Additionally, we no longer sleep waiting for release, which was blocking
the GPIO thread and with the new code would block our main loop, but
instead use timestamps.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>